Update editinvoice command to match invoice.json
