+++ politeiawww/cmswww.go
@@ -139,7 +139,7 @@ func (p *politeiawww) handleUserInvoices(w http.ResponseWriter, r *http.Request)
 	user, err := p.getSessionUser(w, r)
 	if err != nil {
 		RespondWithError(w, r, 0,
-			"handleNewInvoice: getSessionUser %v", err)
+			"handleUserInvoices: getSessionUser %v", err)
 		return
 	}
 
@@ -153,14 +153,14 @@ func (p *politeiawww) handleUserInvoices(w http.ResponseWriter, r *http.Request)
 	util.RespondWithJSON(w, http.StatusOK, reply)
 }
 
-// handleAdminInvoices handles the request to get all of the  of a new contractor by an
-// administrator for the Contractor Management System.
-func (p *politeiawww) handleAdminInvoices(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleAdminInvoices")
-	var ai cms.AdminInvoices
+// handleSetInvoiceStatus handles the incoming set invoice status command.
+func (p *politeiawww) handleSetInvoiceStatus(w http.ResponseWriter, r *http.Request) {
+	// Add the path param to the struct.
+	log.Tracef("handleSetInvoiceStatus")
+	var sis cms.SetInvoiceStatus
 	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&ai); err != nil {
-		RespondWithError(w, r, 0, "handleNewInvoice: unmarshal", www.UserError{
+	if err := decoder.Decode(&sis); err != nil {
+		RespondWithError(w, r, 0, "handleSetInvoiceStatus: unmarshal", www.UserError{
 			ErrorCode: www.ErrorStatusInvalidInput,
 		})
 		return
@@ -169,7 +169,38 @@ func (p *politeiawww) handleAdminInvoices(w http.ResponseWriter, r *http.Request
 	user, err := p.getSessionUser(w, r)
 	if err != nil {
 		RespondWithError(w, r, 0,
-			"handleNewInvoice: getSessionUser %v", err)
+			"handleSetInvoiceStatus: getSessionUser %v", err)
+		return
+	}
+
+	reply, err := p.processSetInvoiceStatus(sis, user)
+	if err != nil {
+		RespondWithError(w, r, 0,
+			"handleSetInvoiceStatus: processSetInvoiceStatus %v", err)
+		return
+	}
+
+	// Reply with the challenge response and censorship token.
+	util.RespondWithJSON(w, http.StatusOK, reply)
+}
+
+// handleAdminInvoices handles the request to get all of the  of a new contractor by an
+// administrator for the Contractor Management System.
+func (p *politeiawww) handleAdminInvoices(w http.ResponseWriter, r *http.Request) {
+	log.Tracef("handleAdminInvoices")
+	var ai cms.AdminInvoices
+	decoder := json.NewDecoder(r.Body)
+	if err := decoder.Decode(&ai); err != nil {
+		RespondWithError(w, r, 0, "handleAdminInvoices: unmarshal", www.UserError{
+			ErrorCode: www.ErrorStatusInvalidInput,
+		})
+		return
+	}
+
+	user, err := p.getSessionUser(w, r)
+	if err != nil {
+		RespondWithError(w, r, 0,
+			"handleAdminInvoices: getSessionUser %v", err)
 		return
 	}
 
@@ -181,6 +212,37 @@ func (p *politeiawww) handleAdminInvoices(w http.ResponseWriter, r *http.Request
 
 	// Reply with the verification token.
 	util.RespondWithJSON(w, http.StatusOK, reply)
+}
+
+// handleEditInvoice attempts to edit an invoice
+func (p *politeiawww) handleEditInvoice(w http.ResponseWriter, r *http.Request) {
+	var ei cms.EditInvoice
+	decoder := json.NewDecoder(r.Body)
+	if err := decoder.Decode(&ei); err != nil {
+		RespondWithError(w, r, 0, "handleEditInvoice: unmarshal",
+			www.UserError{
+				ErrorCode: www.ErrorStatusInvalidInput,
+			})
+		return
+	}
+
+	user, err := p.getSessionUser(w, r)
+	if err != nil {
+		RespondWithError(w, r, 0,
+			"handleEditInvoice: getSessionUser %v", err)
+		return
+	}
+
+	log.Debugf("handleEditInvoice: %v", ei.Token)
+
+	epr, err := p.processEditInvoice(ei, user)
+	if err != nil {
+		RespondWithError(w, r, 0,
+			"handleEditInvoice: processEditInvoice %v", err)
+		return
+	}
+
+	util.RespondWithJSON(w, http.StatusOK, epr)
 }
 
 func (p *politeiawww) setCMSWWWRoutes() {
